test: render watch
